+++ client/client.go
@@ -137,10 +137,8 @@ type Option struct {
 	RPCPath string
 	//ConnectTimeout sets timeout for dialing
 	ConnectTimeout time.Duration
-	// ReadTimeout sets readdeadline for underlying net.Conns
-	ReadTimeout time.Duration
-	// WriteTimeout sets writedeadline for underlying net.Conns
-	WriteTimeout time.Duration
+	// ReadTimeout sets max idle time for underlying net.Conns
+	IdleTimeout time.Duration
 
 	// BackupLatency is used for Failbackup mode. rpcx will sends another request if the first response doesn't return in BackupLatency time.
 	BackupLatency time.Duration
@@ -566,8 +564,8 @@ func (client *Client) send(ctx context.Context, call *Call) {
 		}
 	}
 
-	if client.option.WriteTimeout != 0 {
-		client.Conn.SetWriteDeadline(time.Now().Add(client.option.WriteTimeout))
+	if client.option.IdleTimeout != 0 {
+		client.Conn.SetDeadline(time.Now().Add(client.option.IdleTimeout))
 	}
 
 }
@@ -577,8 +575,8 @@ func (client *Client) input() {
 
 	for err == nil {
 		var res = protocol.NewMessage()
-		if client.option.ReadTimeout != 0 {
-			client.Conn.SetReadDeadline(time.Now().Add(client.option.ReadTimeout))
+		if client.option.IdleTimeout != 0 {
+			client.Conn.SetDeadline(time.Now().Add(client.option.IdleTimeout))
 		}
 
 		err = res.Decode(client.r)